Added comments to process_grism.
Separated the scripts to produce the webpages and related output from the main process_grism script.

New implementation of the direct thumbnails so that they line up with the 2D spectra.

Add FITS images of the direct images and 2D spectra to the output web products.

Added Vizier link (magnifying glass) to the google map.  When clicked, takes you to a Vizier catalog search within 1" of the center position of the map.