Fixed a bug in the WCS alignment.  The shifts had been calculated in the DRZ image (north up) and were being applied to the FLT images (north according to orient).  Now the shifts are rotated into the FLT frame and the alignment to the ACS images looks quite good.
Eventually would like to update the FLT headers directly with the shift information, but this wasn't working very well.  I tried just adding the shifts to the CRPIX keywords in the FLT files, but Multidrizzle would change them back again (with updatewcs=yes).  Multidrizzle would work OK with (updatewcs=no), but I couldn't figure out exactly what it was doing so I left the algorithm as getting the shifts from a shiftfile.